% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_data}
\alias{fetch_data}
\title{Fetch UCI data files}
\usage{
fetch_data(uci_dataset_name, dest = ".")
}
\arguments{
\item{uci_dataset_name}{Name of UCI dataset. See fetchUCI::uci_metadata$name.}

\item{dest}{Directory in which to create a folder. Default to "./".}
}
\value{
A list of results from code{curl::curl_fetch_disk}.
}
\description{
Fetch UCI data files
}
\examples{
\dontrun{
fetch_data("Abalone")
}
}
