% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_description}
\alias{fetch_description}
\title{Fetch UCI data description}
\usage{
fetch_description(uci_dataset_name, dest = ".")
}
\arguments{
\item{uci_dataset_name}{Name of UCI dataset. See fetchUCI::uci_metadata$name.}

\item{dest}{Directory in which to create a folder. Default to "./".}
}
\description{
Fetch UCI data description
}
\examples{
\dontrun{
fetch_description("Abalone")
}
}
